+++ SP Project/GameState.h
@@ -347,17 +347,17 @@ private:
 
 	multimap<float, pointr> dynamic_rendering;
 	multimap<float, light> light_sources;
-	float light_level = 0.1, constant_light_level = 0.3, day_increment = 0.01;
+	float light_level = 0.1f, constant_light_level = 0.3f, day_increment = 0.01f;
 	bool DoDayLightCycle = 1, DoEntitySpawning = 1, EnableMiniMap = 1;
 
 	int size_x = 0, size_y = 0;  //<-- map size
-	int x = 0, y = 0;    //<-- location of upper left corner of the map
+	float x = 0, y = 0;    //<-- location of upper left corner of the map
 	float x_offset = 0, y_offset = 0; //<-- offset from upper left corner of the screen to upper left corner of the map
 	short destructable_count = 0;
 	Vector2i destroy_object_location = { -1, -1 };
 	int object_height = 0;
 	float map_x = 0, map_y = 0;
-	float scale = 1, x_scale = 1, y_scale = 1, win_x = 0, win_y = 0, torch_intensity = 0.6, torch_delta = 0.2;
+	float scale = 1, x_scale = 1, y_scale = 1, win_x = 0, win_y = 0, torch_intensity = 0.6f, torch_delta = 0.2f;
 	RectangleShape blackscreen;
 	Vector2u prev_win = { 0, 0 };
 	Texture minimap_tex;
